ResearchHub-AI/researchpilot/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.22
sentence-transformers==2.2.2
chromadb==0.4.24
pypdf==3.17.1
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
groq>=1.0.0
